requirements-dev.txt
amqp==5.1.1
appdirs==1.4.4
asgiref==3.5.1
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
beautifulsoup4==4.8.0
billiard==3.6.4.0
bs4==0.0.1
cached-property==1.5.2
cattrs==1.10.0
celery==5.2.6
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.9
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography
Deprecated==1.2.13
distlib==0.3.4
Django==3.2.13
django-celery-results==2.3.1        
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
et-xmlfile==1.1.0
filelock==3.6.0
greenlet==1.1.2
h11==0.13.0
idna==2.10
importlib-metadata==4.11.3
kombu==5.2.4
librabbitmq==2.0.0
MechanicalSoup==0.11.0
moodlepy==0.23.5
msgpack==1.0.3
numpy==1.21.5
openpyxl==3.0.9
outcome==1.1.0
packaging==20.4
pandas==1.3.5
platformdirs==2.5.1
plyer==2.0.0
prompt-toolkit==3.0.29
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL
pyparsing==2.4.7
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2022.1
redis==4.2.2
requests==2.27.1
selenium==3.141.0
six==1.15.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.0
spotipy==2.19.0
SQLAlchemy==1.4.36
sqlparse==0.4.2
tqdm==4.64.0
trio==0.20.0
trio-websocket==0.9.2
typing-extensions==4.1.1
urllib3==1.26.9
vine==5.0.0
virtualenv==20.14.0
wcwidth==0.2.5
wrapt==1.14.1
wsproto==1.1.0
zipp==3.7.0
django-rest-swagger
drf-yasg
django-celery-beat
python-decouple
dj-database-url
dj-static